requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
boto3==1.35.0
python-dotenv==1.0.1
langchain==0.3.15
langchain-community==0.3.15
langchain-openai==0.3.6
langchain-text-splitters==0.3.5
pypdf==5.1.0
faiss-cpu==1.9.0.post1